+++ docs/changelog.rst
@@ -76,6 +76,8 @@ Detailed list of changes
 
 - Wayland: Fix specifying the output name for the panel kitten not working (:iss:`7573`)
 
+- icat kitten: Add an option :option:`kitty +kitten icat --no-trailing-newline` to leave the cursor to the right of the image (:iss:`7574`)
+
 
 0.35.2 [2024-06-22]
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~